server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bcrypt>=4.2.1",
    "fastapi>=0.115.6",
    "httpx>=0.28.1",
    "pydantic[email]>=2.10.5",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.0.1",
    "sqlmodel>=0.0.22",
    "uvicorn>=0.34.0",
]
